+++ pyxmex/parser.py
@@ -9,7 +9,7 @@ class Parser(object):
                 content = f.readlines()
                 return self.process_lines(content)
         except UnicodeDecodeError:
-            with open(file_name, 'rb') as f:
+            with open(file_name, encoding='latin-1') as f:
                 content = f.readlines()
                 return self.process_lines(content)